Move functions due to import error
