+++ src/vivarium_gates_mncnh/data/utilities.py
@@ -1,9 +1,7 @@
 from typing import Union
 
-import numpy as np
 import pandas as pd
-from gbd_mapping import ModelableEntity, causes, covariates, risk_factors
-from scipy import stats
+from gbd_mapping import causes, covariates, risk_factors
 from vivarium.framework.artifact import EntityKey
 from vivarium.framework.randomness import get_hash
 from vivarium_inputs.mapping_extension import alternative_risk_factors
@@ -75,67 +73,6 @@ def parse_short_gestation_description(description: str) -> pd.Interval:
         closed="left",
     )
     return endpoints
-
-
-def get_uniform_distribution_from_limits(lower_limit, upper_limit) -> stats.uniform:
-    # stats.uniform is over [loc, loc+scal]
-    loc = lower_limit
-    scale = upper_limit - lower_limit
-    return stats.uniform(loc=loc, scale=scale)
-
-
-def get_norm_from_quantiles(
-    mean: float, lower: float, upper: float, quantiles: tuple[float, float] = (0.025, 0.975)
-) -> stats.norm:
-    stdnorm_quantiles = stats.norm.ppf(quantiles)
-    sd = (upper - lower) / (stdnorm_quantiles[1] - stdnorm_quantiles[0])
-    return stats.norm(loc=mean, scale=sd)
-
-
-def get_norm(
-    mean: float,
-    sd: float = None,
-    ninety_five_pct_confidence_interval: tuple[float, float] = None,
-) -> stats.norm:
-    sd = _get_standard_deviation(mean, sd, ninety_five_pct_confidence_interval)
-    return stats.norm(loc=mean, scale=sd)
-
-
-def get_random_variable_draws(columns: pd.Index, seed: str, distribution) -> pd.Series:
-    return pd.Series(
-        [get_random_variable(x, seed, distribution) for x in range(0, columns.size)],
-        index=columns,
-    )
-
-
-def get_random_variable(draw: int, seed: str, distribution) -> pd.Series:
-    np.random.seed(get_hash(f"{seed}_draw_{draw}"))
-    return distribution.rvs()
-
-
-def _get_standard_deviation(
-    mean: float, sd: float, ninety_five_pct_confidence_interval: tuple[float, float]
-) -> float:
-    if sd is None and ninety_five_pct_confidence_interval is None:
-        raise ValueError(
-            "Must provide either a standard deviation or a 95% confidence interval."
-        )
-    if sd is not None and ninety_five_pct_confidence_interval is not None:
-        raise ValueError(
-            "Cannot provide both a standard deviation and a 95% confidence interval."
-        )
-    if ninety_five_pct_confidence_interval is not None:
-        lower = ninety_five_pct_confidence_interval[0]
-        upper = ninety_five_pct_confidence_interval[1]
-        if not (lower <= mean <= upper):
-            raise ValueError(
-                f"The mean ({mean}) must be between the lower ({lower}) and upper ({upper}) "
-                "quantile values."
-            )
-
-        stdnorm_quantiles = stats.norm.ppf((0.025, 0.975))
-        sd = (upper - lower) / (stdnorm_quantiles[1] - stdnorm_quantiles[0])
-    return sd
 
 
 def set_non_neonnatal_values(data: pd.DataFrame, value: float) -> pd.DataFrame:
